rename extension from aws-replicator to aws-proxy
