+++ aws-proxy/aws_proxy/__init__.py
@@ -1,0 +1,1 @@
+name = "aws-proxy"
