+++ aws-proxy/aws_proxy/client/resource_types.json
