+++ aws-proxy/aws_proxy/server/resource_replicator.py
