+++ aws-proxy/aws_proxy/server/ui/app.js
