+++ aws-proxy/aws_proxy/server/ui/favicon.png
