+++ aws-proxy/aws_proxy/shared/constants.py
