+++ aws-proxy/aws_proxy/shared/utils.py
