+++ aws-proxy/etc/proxy-settings.png
