+++ aws-proxy/example/Makefile
