+++ aws-proxy/example/main.tf
