+++ aws-proxy/setup.cfg
@@ -1,11 +1,11 @@
 [metadata]
-name = localstack-extension-aws-replicator
+name = localstack-extension-aws-proxy
 version = 0.1.25
 summary = LocalStack AWS Proxy Extension
 description = Proxy AWS resources into your LocalStack instance
 long_description = file: README.md
 long_description_content_type = text/markdown; charset=UTF-8
-url = https://github.com/localstack/localstack-extensions/tree/main/aws-replicator
+url = https://github.com/localstack/localstack-extensions/tree/main/aws-proxy
 author = LocalStack Team
 author_email = info@localstack.cloud
 
@@ -44,13 +44,13 @@ test =
     rolo
 
 [options.package_data]
-aws_replicator =
+aws_proxy =
     **/*.html
     **/*.js
     **/*.png
 
 [options.entry_points]
 localstack.extensions =
-    aws-replicator = aws_replicator.server.extension:AwsReplicatorExtension
+    aws-proxy = aws_proxy.server.extension:AwsReplicatorExtension
 localstack.plugins.cli =
-    aws-replicator = aws_replicator.client.cli:AwsReplicatorPlugin
+    aws-proxy = aws_proxy.client.cli:AwsReplicatorPlugin
