--- aws-replicator/MANIFEST.in
@@ -1,3 +1,0 @@
-recursive-include aws_replicator *.html
-recursive-include aws_replicator *.js
-recursive-include aws_replicator *.png
