--- aws-replicator/aws_replicator/__init__.py
@@ -1,1 +1,0 @@
-name = "aws-replicator"
